+++ Sources/PaginatedList.cs
@@ -4,7 +4,12 @@ namespace Mc2it.Agicap;
 /// Represents a strongly typed list of objects, with information relevant to the pagination of its items.
 /// </summary>
 /// <typeparam name="T">The type of items in the list.</typeparam>
-public class PaginatedList<T>: NestedList<T> {
+public class PaginatedList<T> {
+
+	/// <summary>
+	/// The list items.
+	/// </summary>
+	public IList<T> Items { get; set; } = [];
 
 	/// <summary>
 	/// The information relevant to the pagination of list items.